+++ internal/north/rest/router.go
@@ -965,7 +965,9 @@ func NewRouter(d Deps) *chi.Mux { //nolint:gocognit,gocyclo,funlen // compositio
 				pr.With(op).Post("/alarm-messages/ack-all", handlers.AckAllAlarmMessages(d.Hub))
 				pr.With(op).Post("/alarm-messages/{id}/ack", handlers.AckAlarmMessage(d.Hub))
 				pr.Get("/service-messages", handlers.ListServiceMessages(d.Hub))
+				pr.Get("/service-messages/suppressed", handlers.ListSuppressedServiceMessages(d.Hub))
 				pr.With(op).Post("/service-messages/ack-all", handlers.AckAllServiceMessages(d.Hub))
+				pr.With(op).Post("/service-messages/unsuppress", handlers.UnsuppressServiceMessage(d.Hub))
 				pr.With(op).Post("/service-messages/{id}/ack", handlers.AckServiceMessage(d.Hub))
 				pr.With(op).Post("/service-messages/{id}/disable", handlers.DisableServiceMessage(d.Hub))
 			}